+++ src/components/Icon/Icon.tsx
@@ -1,13 +1,47 @@
-import React from 'react'
+import React, { CSSProperties } from 'react'
 import PropTypes from 'prop-types'
 import { customPropTypes, UIComponent, SUI, createShorthandFactory } from '../../lib'
 
 import iconRules from './iconRules'
 import iconVariables from './iconVariables'
 
-class Icon extends UIComponent<any, any> {
-  static create: Function
+export type IconColor =
+  | 'white'
+  | 'red'
+  | 'orange'
+  | 'yellow'
+  | 'olive'
+  | 'green'
+  | 'teal'
+  | 'blue'
+  | 'violet'
+  | 'purple'
+  | 'pink'
+  | 'brown'
+  | 'grey'
+  | 'black'
 
+export type IconSize = 'mini' | 'tiny' | 'small' | 'large' | 'big' | 'huge' | 'massive'
+
+export type IconXSpacing = 'none' | 'before' | 'after' | 'both'
+
+export interface IconProps {
+  as?: string
+  bordered?: boolean
+  circular?: boolean
+  className?: string
+  color?: IconColor
+  disabled?: boolean
+  kind?: string
+  name?: string
+  size?: IconSize
+  xSpacing?: IconXSpacing
+  style?: CSSProperties
+  title?: string
+}
+
+class Icon extends UIComponent<IconProps, {}> {
+  static create: Function
   static className = 'ui-icon'
 
   static displayName = 'Icon'
@@ -56,6 +90,9 @@ class Icon extends UIComponent<any, any> {
 
     /** Size of the icon. */
     size: PropTypes.oneOf(['mini', 'tiny', 'small', 'large', 'big', 'huge', 'massive']),
+
+    /** Adds space to the before, after or on both sides of the icon, or removes the default space around the icon ('none' value) */
+    xSpacing: PropTypes.oneOf(['none', 'before', 'after', 'both']),
   }
 
   static handledProps = [
@@ -68,6 +105,7 @@ class Icon extends UIComponent<any, any> {
     'kind',
     'name',
     'size',
+    'xSpacing',
   ]
 
   static defaultProps = {